requirements.txt
pyserial>=3.5
python-dotenv>=0.19.0
matplotlib>=3.5.0
numpy>=1.21.0
scipy>=1.7.0